Added Dockerfile and adjusted app version.  One small fix.

+++ src/test/java/net/kbg/datamakerri/model/TableTest.java
@@ -337,7 +337,7 @@ public class TableTest {
 
     private Field makeLongField() {
         return Field.builder()
-                .name("DOUBLE")
+                .name("LONG")
                 .dmSourceType(Definitions.LONG)
                 .rangeLowEnd(-20)
                 .rangeHighEnd(20)
